Move Todo functions out of the PostgresDriver.
